# hw_2/src/main/resources/openapi/products.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0

paths:
  /products:
    post:
      summary: Create product
      operationId: createProduct
      tags: [Products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/Error400'

    get:
      summary: List products with pagination and filtering
      operationId: listProducts
      tags: [Products]
      parameters:
        - name: page
          in: query
          description: Page number (starts from 0)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Page size
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
        - name: status
          in: query
          description: Filter by status
          required: false
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, ARCHIVED]
        - name: category
          in: query
          description: Filter by category (exact match)
          required: false
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Page of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPageResponse'
        '400':
          $ref: '#/components/responses/Error400'

  /products/{id}:
    get:
      summary: Get product by ID
      operationId: getProductById
      tags: [Products]
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          $ref: '#/components/responses/Error404Product'

    put:
      summary: Update product
      operationId: updateProduct
      tags: [Products]
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '404':
          $ref: '#/components/responses/Error404Product'

    delete:
      summary: Soft delete product (set status ARCHIVED)
      operationId: archiveProduct
      tags: [Products]
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '204':
          description: Archived (status set to ARCHIVED)
        '404':
          $ref: '#/components/responses/Error404Product'

components:
  parameters:
    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    Error400:
      description: Bad Request (VALIDATION_ERROR)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Error403:
      description: Forbidden (ORDER_OWNERSHIP_VIOLATION)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Error404Product:
      description: Not Found (PRODUCT_NOT_FOUND)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Error409:
      description: Conflict (business rule violation)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Error422:
      description: Unprocessable Entity (promo issues)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Error429:
      description: Too Many Requests (ORDER_LIMIT_EXCEEDED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    ErrorCode:
      type: string
      enum:
        - PRODUCT_NOT_FOUND
        - PRODUCT_INACTIVE
        - ORDER_NOT_FOUND
        - ORDER_LIMIT_EXCEEDED
        - ORDER_HAS_ACTIVE
        - INVALID_STATE_TRANSITION
        - INSUFFICIENT_STOCK
        - PROMO_CODE_INVALID
        - PROMO_CODE_MIN_AMOUNT
        - ORDER_OWNERSHIP_VIOLATION
        - VALIDATION_ERROR

    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true

    ProductStatus:
      type: string
      enum: [ACTIVE, INACTIVE, ARCHIVED]

    ProductCreate:
      type: object
      required: [name, price, stock, category, status]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        price:
          type: string
          format: decimal
          description: Price (> 0)
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
        status:
          $ref: '#/components/schemas/ProductStatus'

    ProductUpdate:
      type: object
      required: [name, price, stock, category, status]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        price:
          type: string
          format: decimal
          description: Price (> 0)
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
        status:
          $ref: '#/components/schemas/ProductStatus'

    ProductResponse:
      type: object
      required: [id, name, price, stock, category, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: string
          format: decimal
        stock:
          type: integer
        category:
          type: string
        status:
          $ref: '#/components/schemas/ProductStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ProductPageResponse:
      type: object
      required: [items, totalElements, page, size]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        totalElements:
          type: integer
          format: int64
          minimum: 0
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1